Better client interaction.
The full safety rules are now implemented including the one that
states: in order for an entry to be considered committed, at least one
entry from the leader's current term must also be stored on a majority
of the servers. In order to make this work with read-only immediate
commands, when a new leader is elected, the leader adds a non-command
entry to the log in order to force previous commands to be committed.
Otherwise, a read might get data before a write from that same client
if the original leader crashed after the write.

Commands now must be a map/object. If the map contains a 'ro' key then
the command is assumed to be a read-only operation and is not added to
the log but rather "applied" immediately to the current state
(committed) of the stateMachine and the callback is invoked
immediately to return the results. Other the commands are still added
to the log and the response to the client is asynchronous when the
command is committed as before.

TODO: idempotent client commands so that commands don't get executed
multiple times if a leader crashes.