Initial addServer/removeServer RPCs
Refactor to have separate internal serverMap from the external
serverData. The internal serverMap represents the members of the
cluster. The serverData in the options map is really information about
how to contact the servers whether they are in the cluster or not.

Add verbose option for control debug level.

Update tests to work with the new model where a single server starts
off and others are added one at a time. The main complication is when
there is stored durable state, then the server assumes it is already
part of the cluster.